# openapi.yaml
openapi: 3.0.3
info:
  title: ias_stt — Trap Management API
  version: 1.0.0
  description: |
    Trap management service with deployment tracking, location
    history, and immutable deployment actions with required pictures. Telemetry is ingested from the
    ChirpStack HTTP integration via `POST /api/telemetry/ingest`
    (an optional MQTT listener remains available). Application users obtain
    JWT access and refresh tokens from `/auth/login` and send access tokens in
    the `Authorization` header. A legacy API key remains available for service
    integrations.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List notifications for the current user
      description: |
        Returns non-expired notifications joined with the current user's
        read and dismissal state. The default `active` status excludes
        notifications dismissed by the current user.
      operationId: listNotifications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, unread, read, dismissed, all]
            default: active
          description: Filter by the current user's notification state
        - $ref: '#/components/parameters/notification_limit'
        - $ref: '#/components/parameters/offset'
        - name: event_type
          in: query
          schema: { type: string, maxLength: 100 }
          description: Filter by event type
        - name: severity
          in: query
          schema: { type: string, maxLength: 20 }
          description: Filter by severity
        - name: entity_type
          in: query
          schema: { type: string, maxLength: 50 }
          description: Filter by related entity type
      responses:
        200:
          description: Notifications and current-user state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        422:
          $ref: '#/components/responses/422'
  /api/notifications/test:
    post:
      tags: [Notifications]
      summary: Create an administrator test notification
      description: |
        Creates a normal in-app notification event for frontend and delivery
        testing. This endpoint requires the administrator `settings:admin`
        permission. The notification is visible to users through the normal
        notification list endpoint and is not sent through external channels.
      operationId: createTestNotification
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestNotificationRequest'
            example:
              title: Test notification
              message: This is a test notification.
              severity: info
      responses:
        201:
          description: Test notification created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        422:
          $ref: '#/components/responses/422'
        500:
          $ref: '#/components/responses/500'
  /api/notifications/summary:
    get:
      tags: [Notifications]
      summary: Get notification counts for the current user
      operationId: notificationSummary
      responses:
        200:
          description: Active and unread notification counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSummary'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
  /api/notifications/{notification_id}:
    parameters:
      - name: notification_id
        in: path
        required: true
        schema: { type: integer }
        description: Notification primary key ID
    get:
      tags: [Notifications]
      summary: Get one notification for the current user
      operationId: getNotification
      responses:
        200:
          description: Notification and current-user state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
  /api/notifications/{notification_id}/read:
    parameters:
      - name: notification_id
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Notifications]
      summary: Mark a notification as read for the current user
      operationId: markNotificationRead
      responses:
        200:
          description: Updated notification state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'
  /api/notifications/{notification_id}/dismiss:
    parameters:
      - name: notification_id
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Notifications]
      summary: Dismiss a notification for the current user
      operationId: dismissNotification
      responses:
        200:
          description: Updated notification state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        404:
          $ref: '#/components/responses/404'

  # -- Trackers ---------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Application users pass the JWT access token as
        `Authorization: Bearer <access-token>`. Legacy service integrations
        may pass the configured API key in the same header.

  # -- Reusable parameters ---------------------------------------------------
  parameters:
...
    notification_limit:
      name: limit
      in: query
      schema: { type: integer, default: 25, minimum: 0, maximum: 100 }
      description: Maximum notifications per page
    offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
      description: Number of items to skip

  # -- Schemas ---------------------------------------------------------------
  schemas:
    Error:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code: { type: string }
                message: { type: string }
      example:
        error:
          code: AUTHENTICATION_REQUIRED
          message: Authentication is required.
...
    Notification:
      type: object
      properties:
        id:            { type: integer, readOnly: true }
        event_type:    { type: string, maxLength: 100 }
        severity:      { type: string, maxLength: 20 }
        title:         { type: string, maxLength: 255 }
        message:       { type: string }
        entity_type:   { type: string, maxLength: 50, nullable: true }
        entity_id:     { type: string, maxLength: 100, nullable: true }
        source_type:   { type: string, maxLength: 50, nullable: true }
        source_id:     { type: string, maxLength: 100, nullable: true }
        payload:
          type: object
          nullable: true
          additionalProperties: true
        dedupe_key:    { type: string, maxLength: 255, nullable: true }
        created_at:    { type: string, format: date-time, readOnly: true }
        expires_at:    { type: string, format: date-time, nullable: true }
        read_at:      { type: string, format: date-time, nullable: true }
        dismissed_at: { type: string, format: date-time, nullable: true }
      required: [id, event_type, severity, title, message, created_at]
    TestNotificationRequest:
      type: object
      additionalProperties: false
      properties:
        title: { type: string, maxLength: 255 }
        message: { type: string, maxLength: 2000 }
        severity: { type: string, enum: [info, warning, critical] }
    NotificationListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Notification' }
        unread_count: { type: integer, minimum: 0 }
        total_count: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 0 }
        offset: { type: integer, minimum: 0 }
      required: [items, unread_count, total_count, limit, offset]
    NotificationSummary:
      type: object
      properties:
        active_count: { type: integer, minimum: 0 }
        unread_count: { type: integer, minimum: 0 }
      required: [active_count, unread_count]
...
  responses:
...
    401:
      description: Missing or invalid JWT or service credential
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: Authentication is required.
    403:
      description: Authenticated user lacks the required permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action.
    422:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Field 'password' must be at least 8 characters
    404:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Trap not found
...
    500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal Server Error
...